refactored cli a bit
